Split signup steps to separate files
