+++ src/app/containers/SignupContainer/PlanStep/EmailSection/EmailSection.js
